requirements.txt
gspread>=5.12.0
google-auth>=2.25.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
requests>=2.31.0
python-dotenv>=1.0.0
lxml>=4.9.3
anthropic>=0.18.0  # Optional: for AI-powered scraping fallback
flask>=3.0.0  # Web interface
